Move tests that deal with encoding directly out of JarIntegrationTest
This way they can be run without non-ASCII characters on the path, so they don't fail with non-Unicode default character encoding.
